refactor: create evm toolkit lib module & cleanup code
