+++ tools/evm-toolkit/src/lib/evm-toolkit/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './network.interface'
