+++ tools/evm-toolkit/src/lib/evm-toolkit/interfaces/network.interface.ts
@@ -1,0 +1,9 @@
+export interface INetwork {
+    id: number;
+    text: string;
+    apiEndpoint: string;
+    faucetEndpoint: string;
+    chainAddress: string;
+    chainID: number;
+    multicallAddress: string;
+}
