--- tools/evm-toolkit/src/lib/network_option.ts
@@ -1,9 +1,0 @@
-export interface NetworkOption {
-  id: number;
-  text: string;
-  apiEndpoint: string;
-  faucetEndpoint: string;
-  chainAddress: string;
-  chainID: number;
-  multicallAddress: string;
-}
